feat: reorganize component structure by moving Carousel, Card, and SuperiorMenu to Areas directory; update imports accordingly;
