+++ components/Areas/SuperiorMenu/style.css
@@ -1,0 +1,171 @@
+@import url('https://fonts.googleapis.com/css2?family=Jacquard+24&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Markazi+Text:wght@400..700&display=swap');
+.superior-menu {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 20px;
+  background-color: #420000;
+}
+
+.brand-area {
+  display: flex;
+  align-items: center;
+}
+
+.brand-area img {
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+  margin-right: 10px;
+}
+
+.brand-area h1 {
+  color: #fff;
+  font-size: 2rem;
+  font-family: "Jacquard 24", serif;
+  font-weight: 400;
+  font-style: normal;
+}
+
+.menu-area {
+  display: flex;
+  list-style: none;
+}
+
+.menu-area ul {
+  display: flex;
+  list-style: none;
+}
+
+.menu-area ul li {
+  margin-right: 20px;
+}
+
+.menu-area ul li a {
+  color: #fff;
+  text-decoration: none;
+  font-size: 1.5rem;
+  font-family: "Markazi Text", serif;
+  font-optical-sizing: auto;
+  font-weight: 400;
+  font-style: normal;
+}
+
+.menu-area ul li a:hover {
+  color: #e9ecef;
+}
+
+.user-area {
+  display: flex;
+  align-items: center;
+}
+
+.user-area img {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  margin-right: 10px;
+}
+
+.user-area h3 {
+  color: #fff;
+  font-size: 1.2rem;
+}
+
+.user-area a {
+  color: #fff;
+  text-decoration: none;
+  font-size: 1.5rem;
+  margin-right: 20px;
+}
+
+.hamburger {
+  display: none;
+  cursor: pointer;
+  font-size: 1.5rem;
+  color: #fff;
+}
+
+.side-menu {
+  display: none;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 250px;
+  height: 100%;
+  background-color: #420000;
+  padding: 20px;
+  box-shadow: 2px 0 5px rgba(0,0,0,0.5);
+  z-index: 1000;
+  flex-direction: column;
+}
+
+.side-menu.active {
+  display: flex;
+}
+
+.side-menu ul {
+  list-style: none;
+  padding: 0;
+}
+
+.side-menu ul li {
+  margin-bottom: 20px;
+}
+
+.side-menu ul li a {
+  color: #fff;
+  text-decoration: none;
+  font-size: 1.2rem;
+}
+
+.side-menu ul li a:hover {
+  color: #e9ecef;
+}
+
+@media (max-width: 768px) {
+  .superior-menu {
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+  }
+  .menu-area {
+    display: none;
+  }
+  .user-area {
+    display: none;
+  }
+  .hamburger {
+    display: block;
+    cursor: pointer;
+    font-size: 1.5rem;
+    color: #fff;
+  }
+  .side-menu.active {
+    display: flex;
+  }
+}
+
+@media (max-width: 480px) {
+  .superior-menu {
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+  }
+  .menu-area {
+    display: none;
+  }
+  .user-area {
+    display: none;
+  }
+  .hamburger {
+    display: block;
+    cursor: pointer;
+    font-size: 1.5rem;
+    color: #fff;
+  }
+  .side-menu.active {
+    display: flex;
+  }
+}
