+++ containers/MoustRecent/index.tsx
@@ -1,4 +1,4 @@
-import Card from "@/components/Card";
+import Card from "@/components/Areas/Card";
 
 const MoustRecent = () => {
     return (
